pyproject.toml
[project]
name = "remote-counter-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "awslabs-mcp-lambda-handler>=0.1.8",
]
